Updated mobile

+++ impilo-mobile/lib/backend/http/sync_service.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:impilo/app_state.dart';
 import 'package:impilo/backend/floor/entities/clinic_data.dart';
 import 'package:impilo/backend/floor/entities/devolve.dart';
+import 'package:impilo/backend/floor/entities/inventory.dart';
 import 'package:impilo/backend/floor/entities/inventory_request.dart';
 import 'package:impilo/backend/floor/entities/refill.dart';
 import 'package:oktoast/oktoast.dart';
@@ -17,41 +18,39 @@ class SyncService {
     List<Devolve> devolves = await _database.devolveDao.findUnSynced();
     List<InventoryRequest> requests =
         await _database.inventoryRequestDao.findUnSynced();
-    List<InventoryRequest> newRequest =
-        requests.where((e) => e.acknowledged == false).toList();
-    List<InventoryRequest> acknowledgements =
-        requests.where((e) => e.acknowledged == true).toList();
+    List<Inventory> acknowledgements =
+        await _database.inventoryDao.findUnSynced();
 
     final Map<String, dynamic> payload = new Map<String, dynamic>();
     payload['refills'] = dispenses.map((d) => d.toJson()).toList();
     payload['clinicData'] = clinic.map((c) => c.toJson()).toList();
     payload['devolves'] = devolves.map((e) => e.toJson()).toList();
-    payload['requests'] = newRequest.map((e) => e.toJson()).toList();
+    payload['requests'] = requests.map((e) => e.toJson()).toList();
     payload['acknowledgements'] = acknowledgements
-        .map((e) => {
-              {
-                'organisation': {'id': FFAppState().code},
-                'uniqueId': e.uniqueId
-              }
-            })
+        .where((e) => e.acknowledged)
+        .map((e) => e.reference)
         .toList();
 
     final response = await api.post(
       '${FFAppState().baseUrl}/api/impilo/mobile-sync',
       data: payload,
     );
+    if (response.data == true) {
+      await _database.clinicDao.updateAllSynced();
+      await _database.devolveDao.updateAllSynced();
+      await _database.refillDao.updateAllSynced();
+      await _database.inventoryRequestDao.updateAllSynced();
+      await _database.inventoryDao.updateAllSynced();
+      var start = DateTime.now().subtract(Duration(days: 366));
 
-    await _database.clinicDao.updateAllSynced();
-    await _database.devolveDao.updateAllSynced();
-    await _database.refillDao.updateAllSynced();
-    await _database.inventoryRequestDao.updateAllSynced();
-    var start = DateTime.now().subtract(Duration(days: 366));
+      await _database.refillDao.deleteOlderThan(start);
+      await _database.devolveDao.deleteAll();
+      await _database.clinicDao.deleteAll();
 
-    await _database.refillDao.deleteOlderThan(start);
-    await _database.devolveDao.deleteAll();
-    await _database.clinicDao.deleteAll();
+      return true;
+    }
 
-    return response.data;
+    return false;
   }
 
   Future<bool> processSync() async {
@@ -65,6 +64,9 @@ class SyncService {
     }
     if (!(hasData ?? false)) {
       hasData = await _database.inventoryRequestDao.hasUnSynced();
+    }
+    if (!(hasData ?? false)) {
+      hasData = await _database.inventoryDao.hasUnSynced();
     }
     if (hasData ?? false) {
       try {
